#50 accontアプリフォルダ構成を整備
