+++ account/forms/__init__.py
@@ -1,0 +1,3 @@
+from .setup_forms import SetupForm
+from .login_forms import LoginForm
+from .signup_forms import SignUpForm
