+++ account/models/__init__.py
@@ -1,0 +1,1 @@
+from .user_models import User
