+++ account/templates/setup.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>initial_setup</title>
+  <title>setup</title>
 </head>
 <body>
   <h1>プロフィール情報</h1>
