+++ account/urls.py
@@ -1,5 +1,5 @@
 from django.urls import path
-from .views import InitialSetupView, SignUpView, LoginView, ProfileView
+from .views import SignUpView, LoginView, SetupView, ProfileView
 from django.contrib.auth.views import LogoutView
 
 
@@ -9,6 +9,6 @@ urlpatterns = [
     path("signup/", SignUpView.as_view(), name="signup"),
     path("login/", LoginView.as_view(), name="login"),
     path("logout/", LogoutView.as_view(), name="logout"),
-    path("initial_setup/", InitialSetupView.as_view(), name="initial_setup"),
+    path("setup/", SetupView.as_view(), name="setup"),
     path("profile/", ProfileView.as_view(), name="profile"),
 ]
