+++ account/views/__init__.py
@@ -1,0 +1,4 @@
+from .login_views import LoginView
+from .signup_views import SignUpView
+from .setup_views import SetupView
+from .profile_views import ProfileView
